Add support to *.zip files.
nats and rclone are now working.
