+++ pkg/get/download.go
@@ -1,9 +1,7 @@
 package get
 
 import (
-	"archive/tar"
 	"bytes"
-	"compress/gzip"
 	"fmt"
 	"github.com/schollz/progressbar/v3"
 	"io"
@@ -11,8 +9,6 @@ import (
 	"net/http"
 	"os"
 	"path"
-	"path/filepath"
-	"strings"
 	"text/template"
 )
 
@@ -164,86 +160,6 @@ func CopyFile(src, dst string, permissions int) (int64, error) {
 		return 0, err
 	}
 	return bites, err
-}
-
-// decompressArchive will decompress any valid archive file.
-func decompressArchive(tool *Tool, dlURL, outFilePath, opSystem, arch, version string) (string, error) {
-	file, err := os.Open(outFilePath)
-	if err != nil {
-		return "", err
-	}
-	defer file.Close()
-
-	outFileDir := filepath.Dir(outFilePath)
-	outFilePath = path.Join(outFileDir, tool.Name)
-
-	switch {
-	case strings.HasSuffix(dlURL, "tar.gz"):
-		err := Untar(file, outFileDir, true)
-		if err != nil {
-			return "", err
-		}
-	case strings.HasSuffix(dlURL, "tgz"):
-		err := Untar(file, outFileDir, true)
-		if err != nil {
-			return "", err
-		}
-	case strings.HasSuffix(dlURL, "zip"):
-		return "", fmt.Errorf("error: zip format not yet supported")
-	default:
-		log.Printf("tool %s appears th have an unsupported binary format\n", tool.Name)
-	}
-	return outFilePath, nil
-}
-
-// Untar untar a file to a target directory on the host it is running on.
-func Untar(r io.Reader, target string, gzip bool) error {
-	return untar(r, target, gzip)
-}
-
-// untar is the private interface for Untar. untar accepts any valid archive
-// and will open it for reading, extract the files and place them into the hosts
-// /tmp directory.
-func untar(r io.Reader, target string, gz bool) error {
-	tarReader := tar.NewReader(r)
-	if gz {
-		gzr, err := gzip.NewReader(r)
-		if err != nil {
-			return fmt.Errorf("failed to decompress gzip: %v", err)
-		}
-		tarReader = tar.NewReader(gzr)
-	}
-	for {
-		header, err := tarReader.Next()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			log.Println("error reading tarfile: %w", err)
-			return fmt.Errorf("tar error: %w", err)
-		}
-
-		fpath := filepath.Base(header.Name)
-		fpath = path.Join(target, fpath)
-		info := header.FileInfo()
-		mode := info.Mode()
-		switch {
-		case mode.IsDir():
-			break
-		case mode.IsRegular():
-			f, err := os.OpenFile(fpath, os.O_RDWR|os.O_CREATE|os.O_TRUNC, mode.Perm())
-			if err != nil {
-				return err
-			}
-			defer f.Close()
-			_, err = io.Copy(f, tarReader)
-			if err != nil {
-				return err
-			}
-		default:
-		}
-	}
-	return nil
 }
 
 type ToolLocal struct {
